frontend: Rename `ApartmentNewSalePage` -> `ApartmentSalePage`
