+++ frontend/src/features/apartment/ApartmentSalePage/ApartmentSalePage.tsx
