+++ frontend/src/features/apartment/ApartmentSalePage/utils.ts
